fix: :bug: compare to date, not separate month and year
Invoiced consultancy contract overview didn't show correct data when the starting date was not in the same year as ending date, because the filter checked separately that month is lower than ending date, even though it can be higher when in another year. This fix replaces the filter with getting a list of PKs by checking for an actual date.